requirements.txt
streamlit==1.28.1
python-pptx==0.6.21
pillow==10.1.0
